ZStream drop and take improvements (#2524)
* Make ZStream#drop take a Long argument

* Check that ZStream#drop is safe to pull again

* Make ZStream#take take a Long attribute

* Make ZStream#take safe to pull again

* Fix rebase issues

* Fix Scala 2.11 type inference issue

+++ streams-tests/jvm/src/test/scala/zio/stream/StreamEffectAsyncSpec.scala
@@ -19,7 +19,7 @@ object StreamEffectAsyncSpec extends ZIOBaseSpec {
           }(global)
         }
 
-        assertM(s.take(list.size).runCollect)(equalTo(list))
+        assertM(s.take(list.size.toLong).runCollect)(equalTo(list))
       })
     ),
     suite("Stream.effectAsyncMaybe")(
@@ -48,7 +48,7 @@ object StreamEffectAsyncSpec extends ZIOBaseSpec {
           None
         }
 
-        assertM(s.take(list.size).runCollect)(equalTo(list))
+        assertM(s.take(list.size.toLong).runCollect)(equalTo(list))
       }),
       testM("effectAsyncMaybe back pressure") {
         for {
@@ -84,7 +84,7 @@ object StreamEffectAsyncSpec extends ZIOBaseSpec {
                       latch.succeed(()) *>
                         Task.unit
                     }
-                    .take(list.size)
+                    .take(list.size.toLong)
                     .run(Sink.collectAll[Int])
                     .fork
           _ <- latch.await
@@ -150,7 +150,7 @@ object StreamEffectAsyncSpec extends ZIOBaseSpec {
           Right(Stream.fromIterable(list))
         }
 
-        assertM(s.take(list.size).runCollect)(equalTo(list))
+        assertM(s.take(list.size.toLong).runCollect)(equalTo(list))
       }),
       testM("effectAsyncInterrupt signal end stream ") {
         for {
